separating gulp into multiple files to improve reading (#1272)
* separating gulp into multiple files to improve reading

* separating the objc part
