+++ scripts/gulp/durations.js
@@ -1,0 +1,35 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2018 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+const _ = require('lodash');
+
+const parseDuration = (duration) => {
+    const ms = parseInt(duration.replace('ms', ''), 10);
+    return ms / 1000;
+  };
+
+const durations = (properties) => _.chain(properties)
+    .filter(({ type }) => type === 'duration')
+    .map(({ value, name, ...rest }) => ({
+        value: parseDuration(value),
+        name,
+        ...rest,
+    }))
+    .value();
+
+module.exports = durations
